requirements.txt
# Django
django==3.2.25  # django
psycopg2-binary==2.8.3  # motor postgres
djangorestframework==3.15.1  # django API


# Testing
pytest==6.2
pytest-django==4.4
pytest-mock==3.6
Faker==9.9.0

# Math
numpy==2.0.1

# Request
requests==2.32.3